Fixed some remarks, added Javadoc and README file

+++ src/main/java/ru/softwerke/practice/app2019/model/Entity.java
@@ -1,5 +1,11 @@
 package ru.softwerke.practice.app2019.model;
 
+/**
+ * Entity common interface
+ */
 public interface Entity {
-    int getId();
+    /**
+     * @return entity's {@code long} identifier
+     */
+    long getId();
 }
